requirements.txt
asgiref==3.8.1
awscli==1.34.32
black==24.10.0
boto3==1.35.32
botocore==1.35.32
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-allauth==65.0.2
django-cloudinary-storage==0.3.0
django-countries==7.6.1
django-crispy-forms==2.3
django-storages==1.14.4
django-widget-tweaks==1.5.0
docutils==0.16
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
stripe==11.1.0
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
